pkg: validator: reorganize to make room for nrt validation
trivial code change/movement to enable adding more
validation, namely the upcoming NRT validation.
