+++ nrovalidate/main.go
@@ -1,18 +1,19 @@
 /*
-Copyright 2022.
+ * Copyright 2022 Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
 package main
 
 import (
@@ -100,12 +101,12 @@ func validateCluster(args ProgArgs) error {
 		return err
 	}
 
-	data, err := nrovalidator.CollectData(context.TODO(), cli)
+	data, err := nrovalidator.Collect(context.TODO(), cli)
 	if err != nil {
 		return err
 	}
 
-	result, err := nrovalidator.Validate(*data)
+	result, err := nrovalidator.Validate(data)
 	if err != nil {
 		return err
 	}
@@ -118,11 +119,11 @@ func validateCluster(args ProgArgs) error {
 
 func printValidationResults(items []deployervalidator.ValidationResult, verbose bool) {
 	if len(items) == 0 {
-		fmt.Printf("PASSED>>: cluster kubelet configuration looks ok!\n")
+		fmt.Printf("PASSED>>: cluster configuration looks ok!\n")
 		return
 	}
 
-	fmt.Printf("FAILED>>: cluster kubelet configuration does NOT look ok!\n")
+	fmt.Printf("FAILED>>: cluster configuration does NOT look ok!\n")
 
 	if !verbose {
 		return
